Updates for sync records containing all double values, and GPS_NMEA_Serial::process() output of doubles. Passes sync_server_dump test.

+++ src/nidas/dynld/GPS_NMEA_Serial.h
@@ -40,10 +40,10 @@ public:
 
 private:
 
-    dsm_time_t parseGGA(const char* input,float *dout,int nvars,dsm_time_t tt) 
+    dsm_time_t parseGGA(const char* input,double *dout,int nvars,dsm_time_t tt) 
         throw();
 
-    dsm_time_t parseRMC(const char* input,float *dout,int nvars,dsm_time_t tt)
+    dsm_time_t parseRMC(const char* input,double *dout,int nvars,dsm_time_t tt)
         throw();
 
     /**
